fix wrong type for poll.h nfds_t
this should not break anything since the type should never be used
except as the argument type for poll.

+++ include/poll.h
@@ -17,7 +17,7 @@ extern "C" {
 #define POLLWRBAND 0x200
 #define POLLMSG    0x400
 
-typedef unsigned int nfds_t;
+typedef unsigned long nfds_t;
 
 struct pollfd
 {
